fix/categories/bugs, add features
Перемістила js файл в папку js.
Додала скролл всіх категорій
Лішки дропдаун лісту ренедеряться через js
Додала курсор поінтер свг з макету
Прибрала баг реагування курсору на закритий дропдаун ліст
Додала ренедер всіх з початку категорій на мобільній версії в кнопку Categories замість Others як на інших версіях
Пофіксила падінги дродаун лісту на моб версії
