+++ src/index.html
@@ -15,8 +15,7 @@
     <include src="./partials/categories.html"></include>
     <include src="./partials/news-cards.html"></include>
 
-
     <script type="module" src="./index.js"></script>
-    <script type="module" src="./partials/categories.js"></script>
+    <script type="module" src="./js/categories.js"></script>
   </body>
 </html>
