+++ src/partials/categories.html
@@ -1,26 +1,26 @@
 <section class="categories">
-  <ul class="categories__list">
+  <ul class="categories__list" id="categories__list">
     <li class="categories__item">
-      <button class="categories__btn display-tab" type="button" id="1"></button>
+      <button class="categories__btn display-tab" type="button"></button>
     </li>
     <li class="categories__item display-tab">
-      <button class="categories__btn" type="button" id="2"></button>
+      <button class="categories__btn" type="button"></button>
     </li>
     <li class="categories__item display-tab">
-      <button class="categories__btn" type="button" id="3"></button>
+      <button class="categories__btn" type="button"></button>
     </li>
     <li class="categories__item display-tab">
-      <button class="categories__btn" type="button" id="4"></button>
+      <button class="categories__btn" type="button"></button>
     </li>
     <li class="categories__item display-desk">
-      <button class="categories__btn" type="button" id="5"></button>
+      <button class="categories__btn" type="button"></button>
     </li>
     <li class="categories__item display-desk">
-      <button class="categories__btn" type="button" id="6"></button>
+      <button class="categories__btn" type="button"></button>
     </li>
     <li class="categories__item">
       <button class="categories__btn" type="button" id="others">
-        <span>Others</span>
+        <span id="categories-span">Others</span>
         <svg class="categories__arrow" width="14" height="14">
           <use href="./images/sprite.svg#arrow-down"></use>
         </svg>
@@ -29,23 +29,6 @@
   </ul>
 
   <div class="dropdown js-is-hidden" id="dropdown">
-    <ul>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li>
-        <button class="dropdown-item" type="button"></button>
-      </li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li>
-        <button class="dropdown-item" type="button"></button>
-      </li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-      <li><button class="dropdown-item" type="button"></button></li>
-    </ul>
+    <ul id="others-list"></ul>
   </div>
 </section>
